Solution to Check Even or Odd Using Division

+++ Check Even or Odd Using Division@@6788dd63490637cfc44e1b9b/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float x;
-    scanf("%f",&x);
+    int x;
+    scanf("%d",&x);
     
     if(x%2==0 && x==2){
         printf("Even");
